backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb~=1.0.13",
    "fake-useragent>=2.2.0",
    "keybert>=0.9.0",
    "llama-index>=0.12.48",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-tools-mcp>=0.2.6",
    "mcp[cli]>=1.10.1",
    "newspaper3k>=0.2.8",
    "ollama~=0.5.1",
    "pandas~=2.3.0",
    "playwright>=1.53.0",
    "pydantic~=2.11.7",
    "python-dotenv~=1.1.0",
    "requests~=2.32.4",
    "sentence-transformers~=4.1.0",
    "torch~=2.7.1",
    "tqdm~=4.67.1",
    "transformers~=4.52.4",
]
